DEV: Autoload app files
