+++ app/services/discourse_translator/base.rb
