+++ app/services/discourse_translator/microsoft.rb
